requirements.txt
Flask
Pillow
numpy
requests
scipy